+++ tools/onnx-graphsurgeon/onnx_graphsurgeon/__init__.py
@@ -6,4 +6,4 @@ from onnx_graphsurgeon.ir.node import Node
 from onnx_graphsurgeon.ir.tensor import Constant, Tensor, Variable
 from onnx_graphsurgeon.util.exception import OnnxGraphSurgeonException
 
-__version__ = "0.4.0"
+__version__ = "0.4.1"